ocr-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
easyocr==1.7.1
opencv-python-headless==4.9.0.80
Pillow==10.3.0
spacy==3.7.4
dateparser==1.2.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
